hbia2/mcp/requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.3
lxml>=4.9.0
aiofiles>=0.8.0
tenacity>=8.0.0 